fix: Move SkillInfo to IPC type binding pipeline
Move SkillInfo struct and new SkillSource enum into solo-protocol
so they go through the ts-rs binding generation pipeline. This
replaces the hand-written TypeScript interface with an auto-generated
one, preventing Rust/TS type drift across the IPC boundary.
